x/fdpd/types/message_calculate_request_delegation_timing_test.go
package types

import (
	"testing"

	"delegationcontrol/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCalculateRequestDelegationTiming_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCalculateRequestDelegationTiming
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCalculateRequestDelegationTiming{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCalculateRequestDelegationTiming{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
